add support to use an external file to specify auto-rebasing branches

+++ .circleci/scripts/auto-rebase.sh
@@ -5,26 +5,37 @@ git config user.email "circleci@bolt.com"
 git config user.name "Circle CI"
 
 baseBranch="master"
-# add the branches that you want to auto-rebase
-# use space to sparate them (don't use comma)
-autoRebaseBranches=("jz/rebase-custom1" "jz/rebase-custom2" "jz/rebase-custom3")
+autoRebaseBranches=()
+# load auto-rebasing branches
+configFile="./.circleci/scripts/auto-rebase-branches.txt"
+if ! test -f "$configFile"; then
+    echo "Cannot find the configuration for auto-rebase."
+    exit 1
+fi
+echo "The following branches will be auto-rebased:"
+while IFS= read -r branchName || [[ -n "$branchName" ]]; do
+  if [ ${#branchName} -gt 0 ]; then
+    autoRebaseBranches+=("$branchName")
+    echo "$branchName"
+  fi
+done < "$configFile"
 
 auto-rebase () {
   local branchName=$1
   printf "\nAuto-rebasing %s to %s...\n" $branchName $baseBranch
   if ! (git checkout "$branchName"); then
     echo "Failed to checkout branch $branchName"
-    exit 1
+    return 1
   fi
   if ! (git rebase origin/$baseBranch); then
     echo "Failed to rebase branch $branchName on $baseBranch"
     git rebase --abort
-    exit 1
+    return 1
   fi
   if ! (git push --force-with-lease origin "$branchName"); then
     echo "Failed to push $branchName"
     git reset --hard HEAD
-    exit 1
+    return 1
   fi
   printf "Auto-rebasing %s to %s is done.\n" $branchName $baseBranch
 }
